+++ beetsplug/discogs.py
@@ -206,13 +206,12 @@ class DiscogsPlugin(BeetsPlugin):
 
         self._log.debug('Searching for release {0}', album_id)
         # Discogs-IDs are simple integers. We only look for those at the end
-        # of an input string as to avoid confusion with other metadata plugins.
+        # of an input string or after the "/release/" keyword as to avoid
+        # confusion with other metadata plugins.
         # An optional bracket can follow the integer, as this is how discogs
         # displays the release ID on its webpage.
-        # Issue #4080 highlighted that Discogs changed the way the URL
-        # is formed
-        match = re.search(r'(^|[*r|discogs\.com/.+/release/])(\d+)($|)',
-                          album_id)
+        match = re.search(
+            r"(^|\[*r|discogs\.com.*/release/)(\d+)(|\])", album_id)
 
         if not match:
             return None